pip-list-requirements.txt
Package                       Version
----------------------------- ---------
astroid                       1.6.5
backports.functools-lru-cache 1.5
certifi                       2018.8.24
chardet                       3.0.4
colorama                      0.3.9
configparser                  3.5.0
Django                        1.7
django-filter                 2.0.0
django-jchart                 0.4.2
enum34                        1.1.6
futures                       3.2.0
idna                          2.7
isort                         4.3.4
lazy-object-proxy             1.3.1
mccabe                        0.6.1
MySQL-python                  1.2.5
passlib                       1.7.1
pbr                           4.2.0
Pillow                        5.2.0
pip                           18.0
pygeocoder                    1.2.5
pylint                        1.9.3
pytz                          2018.5
requests                      2.19.1
setuptools                    39.0.1
singledispatch                3.4.0.3
six                           1.11.0
stevedore                     1.29.0
urllib3                       1.23
virtualenv                    16.0.0
virtualenv-clone              0.3.0
virtualenvwrapper             4.8.2
virtualenvwrapper-win         1.2.5
wheel                         0.31.1
wrapt                         1.10.11